+++ src/Multiformats.Address/Protocols/UDP.cs
@@ -16,19 +16,19 @@ namespace Multiformats.Address.Protocols
         public UDP(uint port)
             : this()
         {
-            Value = port;
+            Value = (int)port;
         }
 
         public UDP(short port)
             : this()
         {
-            Value = port;
+            Value = (int)port;
         }
 
         public UDP(ushort port)
             : this()
         {
-            Value = port;
+            Value = (int)port;
         }
     }
 }
